requirements.txt
# Web framework and server
fastapi>=0.103.1
uvicorn[standard]==0.23.2
jinja2==3.1.2
python-multipart==0.0.6

# HTTP and API
httpx==0.25.0
pyporscheconnectapi>=0.1.0

# Data handling and visualization
pydantic==2.4.2
matplotlib>=3.7.2
Pillow>=10.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
schedule==1.2.0
asyncio>=3.4.3

# CAPTCHA solving
twocaptcha>=0.1.0
cairosvg==2.7.1
